refactor v2 eval special operators
